requirements.txt
# DRF
django==4.2.10
djangorestframework==3.15.1
psycopg[binary]==3.1.12
django-extensions==3.1.3
django-environ==0.11.2
django-filter==24.2

faker==24.4.0

requests==2.31.0
pytz==2024.1

pandas==2.2.2
numpy==1.26.4
xlrd==2.0.1
boto3==1.34.91
gunicorn==22.0.0

# Testing & Coverage
pytest==8.3.3
pytest-django==4.8.0
pytest-cov==5.0.0

# Code formatting & linting
black==24.10.0
isort==5.13.2
flake8==7.1.1
pre-commit==3.8.0
